--- post-type/module-subscriptions-management.php
@@ -1,199 +1,0 @@
-<?php
-if ( !defined( 'ABSPATH' ) ) { exit; } // Exit if accessed directly.
-
-class DT_Subscriptions_Management extends DT_Module_Base {
-    public $module = "subscriptions_management";
-
-    public $magic_link_root = "subscriptions_app"; // define the root of the url {yoursite}/root/type/key/action
-    public $magic_link_type = 'manage'; // define the type
-    public $post_type = 'subscriptions';
-
-    private static $_instance = null;
-    public static function instance() {
-        if ( is_null( self::$_instance ) ) {
-            self::$_instance = new self();
-        }
-        return self::$_instance;
-    } // End instance()
-
-    public function __construct() {
-        parent::__construct();
-        if ( !self::check_enabled_and_prerequisites() ){
-            return;
-        }
-        add_action( 'rest_api_init', [ $this, 'add_api_routes' ] );
-    }
-
-    /**
-     * Register REST Endpoints
-     * @link https://github.com/DiscipleTools/disciple-tools-theme/wiki/Site-to-Site-Link for outside of wordpress authentication
-     */
-    public function add_api_routes() {
-        $namespace = $this->magic_link_root . '/v1';
-        register_rest_route(
-            $namespace, '/'.$this->magic_link_type, [
-                [
-                    'methods'  => "POST",
-                    'callback' => [ $this, 'manage_profile' ],
-                    'permission_callback' => function( WP_REST_Request $request ){
-                        $magic = new DT_Magic_URL( $this->magic_link_root );
-                        return $magic->verify_rest_endpoint_permissions_on_post( $request );
-                    },
-                ],
-            ]
-        );
-        register_rest_route(
-            $namespace, '/'.$this->magic_link_type . '/delete_profile', [
-                [
-                    'methods'  => "DELETE",
-                    'callback' => [ $this, 'delete_profile' ],
-                    'permission_callback' => function( WP_REST_Request $request ){
-                        $magic = new DT_Magic_URL( $this->magic_link_root );
-                        return $magic->verify_rest_endpoint_permissions_on_post( $request );
-                    },
-                ],
-            ]
-        );
-        register_rest_route(
-            $namespace, '/'.$this->magic_link_type . '/allow-notifications', [
-                [
-                    'methods'  => "POST",
-                    'callback' => [ $this, 'allow_notifications' ],
-                    'permission_callback' => function( WP_REST_Request $request ){
-                        $magic = new DT_Magic_URL( $this->magic_link_root );
-                        return $magic->verify_rest_endpoint_permissions_on_post( $request );
-                    },
-                ],
-            ]
-        );
-    }
-
-    public function delete_profile( WP_REST_Request $request ){
-        $params = $request->get_params();
-
-        if ( ! isset( $params['parts'], $params['parts']['meta_key'], $params['parts']['public_key'] ) ) {
-            return new WP_Error( __METHOD__, "Missing parameters", [ 'status' => 400 ] );
-        }
-        $params = dt_recursive_sanitize_array( $params );
-
-        $post_id = $params["parts"]["post_id"]; //has been verified in verify_rest_endpoint_permissions_on_post()
-        if ( ! $post_id ){
-            return new WP_Error( __METHOD__, "Missing post record", [ 'status' => 400 ] );
-        }
-        global $wpdb;
-
-        //remove connection
-        $a = $wpdb->query( $wpdb->prepare( "UPDATE $wpdb->posts p SET post_title = 'Deleted Subscription', post_name = 'Deleted Subscription' WHERE p.ID = %s", $post_id ) );
-        $a = $wpdb->query( $wpdb->prepare( "DELETE FROM $wpdb->dt_activity_log WHERE object_id = %s and action != 'add_subscription' and action !='delete_subscription'", $post_id ) );
-        //create activity for connection removed on the campaign
-        DT_Posts::update_post( 'subscriptions', $post_id, [ "campaigns" => [ "values" => [], "force_values" => true ] ], true, false );
-        $a = $wpdb->query( $wpdb->prepare( "DELETE pm FROM $wpdb->postmeta pm WHERE pm.post_id = %s", $post_id ) );
-        $a = $wpdb->query( $wpdb->prepare( "DELETE FROM $wpdb->dt_reports WHERE post_id = %s", $post_id ) );
-        DT_Posts::update_post( 'subscriptions', $post_id, [ "status" => 'inactive' ], true, false );
-
-        return true;
-    }
-
-    public function manage_profile( WP_REST_Request $request ) {
-        $params = $request->get_params();
-
-        if ( ! isset( $params['parts'], $params['parts']['meta_key'], $params['parts']['public_key'], $params['action'] ) ) {
-            return new WP_Error( __METHOD__, "Missing parameters", [ 'status' => 400 ] );
-        }
-
-        $params = dt_recursive_sanitize_array( $params );
-        $action = sanitize_text_field( wp_unslash( $params['action'] ) );
-
-        // manage
-        $post_id = $params["parts"]["post_id"]; //has been verified in verify_rest_endpoint_permissions_on_post()
-        if ( ! $post_id ){
-            return new WP_Error( __METHOD__, "Missing post record", [ 'status' => 400 ] );
-        }
-
-        switch ( $action ) {
-            case 'get':
-                return $this->get_subscriptions( $post_id );
-            case 'delete':
-                return $this->delete_subscriptions( $post_id, $params );
-            case 'add':
-                return $this->add_subscriptions( $post_id, $params );
-            default:
-                return new WP_Error( __METHOD__, "Missing valid action", [ 'status' => 400 ] );
-        }
-    }
-
-    public function get_subscriptions( $post_id ) {
-        $subs = Disciple_Tools_Reports::get( $post_id, 'post_id' );
-
-        if ( ! empty( $subs ) ){
-            foreach ( $subs as $index => $sub ) {
-                // verification step
-                if ( $sub['value'] < 1 ) {
-                    Disciple_Tools_Reports::update( [
-                        'id' => $sub['id'],
-                        'value' => 1
-                    ] );
-                    $subs[$index]['value'] = 1;
-                    $subs[$index]['verified'] = true;
-                }
-
-                $subs[$index]['formatted_time'] = gmdate( 'F d, Y @ H:i a', $sub['time_begin'] ) . ' for ' . $sub['label'];
-            }
-        }
-
-        return $subs;
-    }
-
-    private function delete_subscriptions( $post_id, $params ) {
-        $sub = Disciple_Tools_Reports::get( $params["report_id"], 'id' );
-        $time_in_mins = ( $sub["time_end"] - $sub["time_begin"] ) / 60;
-        //@todo convert timezone?
-        $label = "Commitment deleted: " . gmdate( 'F d, Y @ H:i a', $sub['time_begin'] ) . ' UTC for ' . $time_in_mins . ' minutes';
-        Disciple_Tools_Reports::delete( $params['report_id'] );
-        dt_activity_insert([
-            'action' => 'delete_subscription',
-            'object_type' => $this->post_type, // If this could be contacts/groups, that would be best
-            'object_subtype' => 'report',
-            'object_note' => $label,
-            'object_id' => $post_id
-        ] );
-        return $this->get_subscriptions( $post_id );
-    }
-
-    private function add_subscriptions( $post_id, $params ){
-        $post = DT_Posts::get_post( 'subscriptions', $post_id, true, false );
-        if ( !isset( $post["campaigns"][0]["ID"] ) ){
-            return false;
-        }
-        $campaign_id = $post["campaigns"][0]["ID"];
-
-        foreach ( $params['selected_times'] as $time ){
-            if ( !isset( $time["time"] ) ){
-                continue;
-            }
-            $new_report = DT_Subscriptions::add_subscriber_time( $campaign_id, $post_id, $time["time"], $time["duration"], $time['grid_id'] );
-            if ( !$new_report ){
-                return new WP_Error( __METHOD__, "Sorry, Something went wrong", [ 'status' => 400 ] );
-            }
-        }
-        return $this->get_subscriptions( $params['parts']['post_id'] );
-    }
-
-
-    public function allow_notifications( WP_REST_Request $request ){
-        $params = $request->get_params();
-
-        if ( ! isset( $params['parts'], $params['parts']['meta_key'], $params['parts']['public_key'] ) ) {
-            return new WP_Error( __METHOD__, "Missing parameters", [ 'status' => 400 ] );
-        }
-        $params = dt_recursive_sanitize_array( $params );
-
-        $post_id = $params["parts"]["post_id"]; //has been verified in verify_rest_endpoint_permissions_on_post()
-        if ( ! $post_id ){
-            return new WP_Error( __METHOD__, "Missing post record", [ 'status' => 400 ] );
-        }
-
-        return update_post_meta( $post_id, "receive_prayer_time_notifications", !empty( $params["allowed"] ) );
-
-    }
-}
